positioning tilted window after untilted at start
Former-commit-id: 2542a7485b3c382f883b86b2dc941fd1793a1766

+++ java/src/xmipp/viewer/particlepicker/ParticlePickerCanvas.java
@@ -144,7 +144,7 @@ public abstract class ParticlePickerCanvas extends XmippImageCanvas
 	}
 
 
-	public void display(float xlocation, float ylocation)
+	public void display(double xlocation, double ylocation)
 	{
 		boolean relocate = (iw == null);
